riorganizzato il package screens
ho creato file appositi per le pagine di login, home e settings
